+++ sdk/search/azure-search-documents/azure/search/documents/models/__init__.py
@@ -25,20 +25,28 @@
 # --------------------------------------------------------------------------
 
 from .._generated.models import (
+    Answers,
+    AnswerResult,
     AutocompleteMode,
     IndexAction,
     IndexingResult,
+    QueryLanguage,
     QueryType,
     SearchMode,
+    Speller,
 )
 from .._search_client import odata
 
 
 __all__ = (
+    "Answers",
+    "AnswerResult",
     "AutocompleteMode",
     "IndexAction",
     "IndexingResult",
     "odata",
+    "QueryLanguage",
     "QueryType",
     "SearchMode",
+    "Speller",
 )