+++ sdk/translation/azure-ai-translation-document/azure/ai/translation/document/_helpers.py
@@ -16,20 +16,24 @@ from ._generated.models import (
     TargetInput as _TargetInput,
 )
 from ._models import DocumentTranslationInput
+
 COGNITIVE_KEY_HEADER = "Ocp-Apim-Subscription-Key"
 POLLING_INTERVAL = 1
 
 
 def get_translation_input(args, kwargs, continuation_token):
     try:
-        inputs = kwargs.pop('inputs', None)
+        inputs = kwargs.pop("inputs", None)
         if not inputs:
             inputs = args[0]
-        request = DocumentTranslationInput._to_generated_list(inputs) \
-            if not continuation_token else None  # pylint: disable=protected-access
+        request = (
+            DocumentTranslationInput._to_generated_list(inputs)  # pylint: disable=protected-access
+            if not continuation_token
+            else None
+        )
     except (AttributeError, TypeError, IndexError):
         try:
-            source_url = kwargs.pop('source_url', None)
+            source_url = kwargs.pop("source_url", None)
             if not source_url:
                 source_url = args[0]
             target_url = kwargs.pop("target_url", None)
@@ -40,18 +44,19 @@ def get_translation_input(args, kwargs, continuation_token):
                 target_language_code = args[2]
             request = [
                 _BatchRequest(
-                    source=_SourceInput(
-                        source_url=source_url
-                    ),
-                    targets=[_TargetInput(
-                        target_url=target_url,
-                        language=target_language_code
-                    )]
+                    source=_SourceInput(source_url=source_url),
+                    targets=[
+                        _TargetInput(
+                            target_url=target_url, language=target_language_code
+                        )
+                    ],
                 )
             ]
         except (AttributeError, TypeError, IndexError):
-            raise ValueError("Pass 'inputs' for multiple inputs or 'source_url', 'target_url', "
-                             "and 'target_language_code' for a single input.")
+            raise ValueError(
+                "Pass 'inputs' for multiple inputs or 'source_url', 'target_url', "
+                "and 'target_language_code' for a single input."
+            )
 
     return request
 
@@ -85,7 +90,7 @@ def get_http_logging_policy(**kwargs):
             "Set-Cookie",
             "X-Powered-By",
             "Strict-Transport-Security",
-            "x-content-type-options"
+            "x-content-type-options",
         }
     )
     http_logging_policy.allowed_query_params.update(
@@ -97,7 +102,7 @@ def get_http_logging_policy(**kwargs):
             "statuses",
             "createdDateTimeUtcStart",
             "createdDateTimeUtcEnd",
-            "$orderBy"
+            "$orderBy",
         }
     )
     return http_logging_policy