+++ sdk/metricsadvisor/azure-ai-metricsadvisor/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Release History
 
+## 1.0.0b5 (Unreleased)
+
+
 ## 1.0.0b4 (2021-06-07)
 
 **New Features**